backend/requirements.txt
Django==3.2.6
flake8==3.9.2
django-colorfield==0.6.3
django-filter==2.4.0
djangorestframework==3.12.4
drf-extra-fields==3.4.0
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
Pillow==9.0.0
python-dotenv==0.19.2

